PMM-7580 Refactor database versions to hook
